chore: 주석으로 로직 설명 추가 및 잘못된 모델함수 이동
